requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pymysql==1.1.0
python-dotenv==1.0.1
jinja2==3.1.3
# Add a barcode library (python-barcode is well-maintained)
python-barcode==0.15.1
# For generating barcode images, Pillow is required
pillow==10.2.0
